+++ react-client/src/components/ProductList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ProductListItem from './ProductListItem.jsx';
-
+import {Route} from 'react-router-dom'
 
 const ProductList = function(props) {
   const listItems = props.items.map((item, index) => {
@@ -9,5 +9,18 @@ const ProductList = function(props) {
 
   return <ul> {listItems} </ul>
 }
+
+
+const ProductDetail = () => {
+  debugger
+  (
+  <Switch>
+    <Route exact path='/products' component={ProductList}/>
+    <Route path='/products/:number' component={Player}/>
+  </Switch>
+)
+}
+
+
 
 export default ProductList;
