+++ react-client/src/components/ProductListItem.jsx
@@ -1,47 +1,72 @@
 import React from 'react';
 import Toggle from 'react-toggle'
 import ProductDetail from './ProductDetail.jsx'
-import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-
-// import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
-// import {Route, NavLink,HashRouter} from "react-router-dom";
-
-
+import { Link,  Route} from "react-router-dom";
 const ProductListItem = (props) => {
 
-  let toggleSwitch;
-  if (props.isLoggedIn) {
-    toggleSwitch = <span> <Toggle /> </span>
-  }
-
-  return (
-    <Router>
-      <div>
-        <li>
-          <Link to=
-          {{ pathname: '/products', search: `?id=${props.value.itemId}` }} > { props.value.name } Price:${props.value.msrp}
-          {toggleSwitch}
-          </Link>
-        </li>
-
-
-
-        <div className="content">
-          <Route exact path='/products' component={ProductDetail}/>
-        </div>
-
-      </div>
-
-
-
-    </Router>
-    )
+let toggleSwitch;
+if (props.isLoggedIn) {
+  toggleSwitch = <span> <Toggle /> </span>
 }
-
+ return (
+  <div>
+    <li>
+      <Link to={`/products/${props.value.itemId}`}> { props.value.name } Price:${props.value.msrp}
+      {toggleSwitch}
+      </Link>
+    </li>
+  </div>
+)}
 
 export default ProductListItem;
 
 
+
+
+//////////////////////////////////////// Attempt 2
+
+// import React from 'react';
+// import Toggle from 'react-toggle'
+// import ProductDetail from './ProductDetail.jsx'
+// import { BrowserRouter as Router, Route, Link } from "react-router-dom";
+
+// const ProductListItem = (props) => {
+
+//   let toggleSwitch;
+//   if (props.isLoggedIn) {
+//     toggleSwitch = <span> <Toggle /> </span>
+//   }
+
+//   props.
+
+//   return (
+//     <Router>
+//       <div>
+//         <li>
+//           <Link to=
+//           {{ pathname: '/products', search: `?id=${props.value.itemId}` }} > { props.value.name } Price:${props.value.msrp}
+//           {toggleSwitch}
+//           </Link>
+//         </li>
+
+
+
+//         <div className="content">
+//           <Route exact path='/products' component={ProductDetail}/>
+//         </div>
+
+//       </div>
+
+
+
+//     </Router>
+//     )
+// }
+
+
+// export default ProductListItem;
+
+//////////////////////////////////////// Attempt 1
 
 // import React from 'react';
 // import Toggle from 'react-toggle'
@@ -58,8 +83,8 @@ export default ProductListItem;
 // }
 //  return (
 
-
-//     <HashRouter>
+//     <H
+// ashRouter>
 //     <div>
 //       <li>
 //         <NavLink to='/products'> { props.value.name } Price:${props.value.msrp}
